requirement.txt
streamlit>=1.35.0
groq>=0.9.0
pillow>=10.3.0
requests>=2.31.0
gTTS>=2.5.0
moviepy>=1.0.3
python-dotenv>=1.0.0